app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["zaynAlii <ziynali938@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
sqlmodel = "^0.0.16"
psycopg = {extras = ["binary"], version = "^3.1.18"}
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
tenacity = "^8.2.3"
pytest = "^8.1.2"
httpx = "^0.27.0"
requests = "^2.31.0"
python-multipart = "^0.0.9"
pydantic-settings = "^2.3.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}




# pin bcrypt passlib supports the latest
bcrypt = "^4.1.3"
pyjwt = "^2.8.0"




[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
